feat: refactor notes API for improved async handling and add test helper functions

+++ notes-backend/eslint.config.mjs
@@ -19,7 +19,7 @@ export default [
     rules: {
       "@stylistic/js/indent": ["error", 2],
       "@stylistic/js/linebreak-style": ["error", "unix"],
-      "@stylistic/js/quotes": ["error", "double"],
+      "@stylistic/js/quotes": ["error", "single"],
       eqeqeq: "error",
       "no-trailing-spaces": "error",
       "object-curly-spacing": ["error", "always"],
